reviewinn-backend/requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0

# Authentication and Security
python-jose[cryptography]==3.4.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# HTTP and API
httpx==0.28.1
requests==2.32.4

# AI Services
google-generativeai==0.8.3

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Caching and Search
redis==5.2.1
elasticsearch==8.17.0

# Background Tasks
celery==5.4.0
flower==2.0.1

# Monitoring and Logging
structlog==25.4.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.19.2

# Development Tools
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3